Move fetch transactions to Events featur module
